[Sanjay] add. saving data into database

+++ src/main/java/com/example/EmployeePayrollApp/repository/EmployeeRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+    // Custom queries can be added here if necessary, like find by name, salary, etc.
 }
